Changed brion::CompartmentReport::loadFrame return type Frame
This is needed to call the function directly from brain::load without
using a deferred future for assembling the Frame struct. The problem
of using a deferred future is that wait_for becomes unusable. Using
a normal async doesn't seem reasonable as it spawns a thread.

Added missing function in CompartmentReportView. The return type was
changed to be by value because by reference was not incorrect.